% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PmcaCovariate.R
\name{createPMCACovariateSettings}
\alias{createPMCACovariateSettings}
\title{Extracts PMCA covariates}
\usage{
createPMCACovariateSettings(
  startDay = -1096,
  endDay = 0,
  analysisId = 763,
  useBulkLoad = T
)
}
\arguments{
\item{startDay}{The number of days prior to index to start observing the cohort}

\item{endDay}{The number of days prior to index to stop observing the cohort}

\item{analysisId}{The analysisId for the covariate}

\item{useBulkLoad}{Whether to use bulk upload for the PMCA variables}
}
\value{
An object of class covariateSettings specifying how to create the cohort covariate with the covariateId
 cohortId x 100000 + settingId x 1000 + analysisId
}
\description{
Extracts PMCA covariates
}
\details{
The user specifies a cohort and time period and then a covariate is constructed whether they are in the
cohort during the time periods relative to target population cohort index
}
