% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisitLocationCovariate.R
\name{createVisitLocationCovariateSettings}
\alias{createVisitLocationCovariateSettings}
\title{Extracts visit location covariate}
\usage{
createVisitLocationCovariateSettings(analysisId = 764)
}
\arguments{
\item{analysisId}{The analysisId for the covariate}
}
\value{
An object of class covariateSettings specifying how to create the VisitLocation covariate
}
\description{
Extracts visit location covariate
}
\details{
Determines the visit type for the cohort index
}
