% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  databaseDetails,
  outputFolder,
  createCohorts = F,
  getData = F,
  packageResults = F,
  minCellCount = 5,
  sampleSize = NULL,
  logSettings,
  useBulkLoad = T
)
}
\arguments{
\item{databaseDetails}{Database details created using \code{PatientLevelPrediction::createDatabaseDetails()}}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the cohortTable table with the target population and outcome cohorts?}

\item{getData}{Extract the patient level data and save as a csv (this is not shared)}

\item{packageResults}{Should results be packaged for later sharing?}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included 
in packaged results.}

\item{sampleSize}{The number of patients in the target cohort to sample (if NULL uses all patients)}

\item{logSettings}{The log setting \code{PatientLevelPrediction::createLogSettings()}}

\item{useBulkLoad}{Use bulk load when inserting a large table into the database}
}
\description{
Execute the Study
}
\details{
This function executes the PASC Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")
                                             
 databaseDetails <- PatientLevelPrediction::createDatabaseDetails(
 connectionDetails = connectionDetails,
 cdmDatabaseSchema = cdmDatabaseSchema,
 cdmDatabaseName = cdmDatabaseName,
 tempEmulationSchema = tempEmulationSchema,
 cohortDatabaseSchema = cohortDatabaseSchema,
 cohortTable = cohortTable,
 outcomeDatabaseSchema = cohortDatabaseSchema,
 outcomeTable = cohortTable,
 cdmVersion = cdmVersion
 )  
 
 logSettings <- PatientLevelPrediction::createLogSettings(
 verbosity = "INFO",
 timeStamp = T,
 logName = 'skeletonPlp'
 )                                          

execute(databaseDetails = databaseDetails,
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        getData = T,
        packageResults = F,
        minCellCount = 5,
        sampleSize = 10000,
        logSettings = logSettings
        )
}

}
